change ColumnPicker to a clean ColumnSelect
